+++ ember_mug/cli/__init__.py
@@ -1,0 +1,5 @@
+"""CLI Interface."""
+
+from .commands import EmberMugCli
+
+__all__ = ('EmberMugCli',)